Creating the reports for fixed time periods. They store the average followers and mentions and also the users sorted by their acceleration

+++ models/user_report.rb
@@ -4,8 +4,8 @@ class UserReport < Hash
 
 	property :time
 	property :mentions, Integer, :default => 0
-	property :followers, Integer, :default => 0
-	property :acceleration, Integer, :default => 0
-	property :velocity, Integer, :default => 0
+	property :followers, Float, :default => 0
+	property :acceleration, Float, :default => 0
+	property :velocity, Float, :default => 0
 
 end
